Issue #17 Refactored the package structure for professor and student modules
