+++ pom.xml
@@ -13,6 +13,11 @@
         <module>student</module>
     </modules>
 
+    <properties>
+        <maven.compiler.source>1.6</maven.compiler.source>
+        <maven.compiler.target>1.6</maven.compiler.target>
+    </properties>
+
     <packaging>pom</packaging>
     
 </project>
