+++ professor/src/main/java/app/ProfessorApp.java
@@ -1,5 +1,8 @@
-import javax.swing.*;
+package main.java.app;
 
+import main.java.quizcreator.ProfessorModel;
+import main.java.quizcreator.ProfessorView;
+import main.java.quizcreator.ProfessorViewController;
 
 
 public class ProfessorApp {
@@ -7,6 +10,5 @@ public class ProfessorApp {
         ProfessorModel professorModel = new ProfessorModel("Quiz1.json");
         ProfessorView professorView = new ProfessorView();
         ProfessorViewController professorViewController = new ProfessorViewController(professorView, professorModel);
-
     }
 }
