+++ professor/src/main/java/quizcreator/ProfessorView.java
@@ -1,3 +1,5 @@
+package main.java.quizcreator;
+
 import com.intellij.uiDesigner.core.GridConstraints;
 import com.intellij.uiDesigner.core.GridLayoutManager;
 import com.intellij.uiDesigner.core.Spacer;
