+++ professor/src/main/java/ui/QuizSpec.java
@@ -1,3 +1,5 @@
+package main.java.ui;
+
 import com.intellij.uiDesigner.core.GridConstraints;
 import com.intellij.uiDesigner.core.GridLayoutManager;
 import com.intellij.uiDesigner.core.Spacer;
