+++ quizcore/pom.xml
@@ -23,6 +23,9 @@
             <version>2.1</version>
         </dependency>
     </dependencies>
+    <build>
+        <sourceDirectory>src</sourceDirectory>
+    </build>
 
 
 </project>
